refactor: Rename ActorsLayout to ActorLazyVerticalColumn and update usage
