+++ presentation/src/main/java/com/london/presentation/feature/home/trending/actor/TrendingActorsScreen.kt
@@ -1,10 +1,10 @@
 package com.london.presentation.feature.home.trending.actor
 
 import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -19,10 +19,10 @@ import androidx.paging.compose.collectAsLazyPagingItems
 import com.london.designsystem.component.TopBar
 import com.london.designsystem.theme.NovixTheme
 import com.london.presentation.R
-import com.london.presentation.shared.ActorItem
-import com.london.presentation.shared.container.LazyPagingColumn
+import com.london.presentation.shared.container.ActorLazyVerticalColumn
 import com.london.presentation.shared.buildscreen.BuildScreen
 import com.london.presentation.utils.Listen
+import com.london.presentation.utils.isLoading
 
 @Composable
 fun TrendingActorsScreen(
@@ -43,21 +43,10 @@ fun TrendingActorsScreen(
         }
     }
 
-    val actorsLazyItems = state.actorsFlow.collectAsLazyPagingItems()
-
-    BuildScreen(
-        isLoading = state.isLoading,
-        isError = actorsLazyItems.loadState.refresh is LoadState.Error,
-        onBack = viewModel::onBackClick,
-        onRetry = viewModel::onRetryClick,
-        emptyLayoutMessage = R.string.no_trending_actors_in_genre,
-        emptyLayoutImage = R.drawable.img_no_result,
-    ) {
-        Content(
-            state = state,
-            contract = viewModel,
-        )
-    }
+    Content(
+        state = state,
+        contract = viewModel
+    )
 }
 
 @Composable
@@ -65,35 +54,34 @@ private fun Content(
     state: TrendingActorsUiState,
     contract: TrendingActorsContract,
 ) {
-    LazyColumn(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(horizontal = 16.dp)
-            .background(color = NovixTheme.colors.surface)
+    val actorsLazyItems = state.actorsFlow.collectAsLazyPagingItems()
+
+    BuildScreen(
+        isLoading = actorsLazyItems.isLoading(),
+        isError = actorsLazyItems.loadState.refresh is LoadState.Error,
+        onBack = contract::onBackClick,
+        onRetry = contract::onRetryClick,
+        emptyLayoutMessage = R.string.no_trending_actors_in_genre,
+        emptyLayoutImage = R.drawable.img_no_result,
+        pagingFlow = actorsLazyItems
     ) {
-        stickyHeader {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(color = NovixTheme.colors.surface)
+        ) {
             TopBar(
-                title = stringResource(R.string.trending_people),
+                title = stringResource(com.london.designsystem.R.string.tv_shows),
                 onBackClick = contract::onBackClick,
                 modifier = Modifier
                     .fillMaxWidth()
                     .background(NovixTheme.colors.surface)
-                    .padding(vertical = 12.dp)
+                    .padding(start = 16.dp, end = 16.dp, bottom = 8.dp)
             )
-        }
 
-        item {
-            LazyPagingColumn(
-                pagingItems = state.actorsFlow.collectAsLazyPagingItems(),
-                modifier = Modifier.fillMaxSize(),
-                itemContent = { actor ->
-                    ActorItem(
-                        actorName = actor.name,
-                        characterName = null,
-                        imageRes = actor.profilePictureUrl,
-                        onClick = { contract.onActorClick(actor.id) }
-                    )
-                }
+            ActorLazyVerticalColumn(
+                items = actorsLazyItems,
+                onActorClick = { contract.onActorClick(it.id) }
             )
         }
     }
@@ -104,6 +92,10 @@ private fun Content(
 private fun Preview() = NovixTheme {
     Content(
         state = TrendingActorsUiState(),
-        contract = defaultTrendingActorsContract()
+        contract = object : TrendingActorsContract {
+            override fun onActorClick(id: Int) {}
+            override fun onBackClick() {}
+            override fun onRetryClick() {}
+        }
     )
 }
