+++ presentation/src/main/java/com/london/presentation/shared/container/ActorLazyVerticalColumn.kt
@@ -1,4 +1,4 @@
-package com.london.presentation.shared
+package com.london.presentation.shared.container
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
@@ -9,16 +9,17 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.paging.compose.LazyPagingItems
 import com.london.domain.entity.actor.Actor
+import com.london.presentation.shared.ActorItem
 
 @Composable
-fun ActorsLayout(
+fun ActorLazyVerticalColumn(
     items: LazyPagingItems<Actor>,
     onActorClick: (Actor) -> Unit,
     modifier: Modifier = Modifier
 ) {
     LazyColumn(
         modifier = modifier.fillMaxWidth(),
-        contentPadding = PaddingValues(start = 16.dp, end = 16.dp, bottom = 16.dp),
+        contentPadding = PaddingValues(start = 16.dp, end = 16.dp, bottom = 16.dp, top = 12.dp),
         verticalArrangement = Arrangement.spacedBy(12.dp)
     ) {
         items(items.itemCount) { index ->
